+++ lib/lightning/collections/collection.ex
@@ -18,6 +18,7 @@ defmodule Lightning.Collections.Collection do
   schema "collections" do
     field :name, :string
     field :byte_size_sum, :integer
+    field :delete, :boolean, virtual: true
     belongs_to :project, Lightning.Projects.Project
     has_many :items, Lightning.Collections.Item
 
@@ -29,6 +30,11 @@ defmodule Lightning.Collections.Collection do
     entry
     |> cast(attrs, [:project_id, :name])
     |> validate_required([:project_id, :name])
+    |> validate()
+  end
+
+  def validate(changeset) do
+    changeset
     |> validate_format(:name, ~r/^[a-z0-9]+([\-_.][a-z0-9]+)*$/,
       message: "Collection name must be URL safe"
     )