+++ lib/lightning/projects/project.ex
@@ -40,6 +40,8 @@ defmodule Lightning.Projects.Project do
 
     has_many :project_credentials, ProjectCredential
     has_many :credentials, through: [:project_credentials, :credential]
+    has_many :collections, Lightning.Collections.Collection
+
     timestamps()
   end